Refactor CLI to use Effect-TS for command handling and file operations
- Replaced existing command handling with Effect-TS Command API.
- Introduced new functions for running shell commands and managing file operations using Effect-TS.
- Implemented comprehensive index monitoring for controller files and root index files.
- Updated package.json generation to use Effect-TS for file reading and writing.
- Enhanced the watcher functionality to monitor API resources and models for changes.
- Removed legacy code and dependencies, streamlining the CLI structure.
